drop_voice_back/requirements.txt
Flask==2.2.5
Werkzeug==2.2.2
python-dotenv==1.0.0
pycryptodome
pyrebase4==4.7.1
requests==2.29.0
firebase-admin==6.2.0
flask-cors==4.0.0
Flask-SQLAlchemy==3.1.1
flask-restx==1.2.0
mysqlclient==2.2.0
flask_sslify==0.1.5
dropbox==11.36.2
gunicorn==20.1.0